Renames typemaps to completion maps
