+++ continuationsPlugin/src/main/scala/continuations/types/DefDefAddCompletionMap.scala
@@ -5,11 +5,10 @@ import dotty.tools.dotc.core.Contexts.Context
 import dotty.tools.dotc.core.Symbols.*
 import dotty.tools.dotc.core.Types.*
 
-final class ValDefAddCompletionType()(using Context) extends TypeMap {
+final class DefDefAddCompletionMap()(using Context) extends TypeMap {
   def hasSuspendClass(s: Symbol)(using ctx: Context): Boolean =
     s.info.hasClassSymbol(requiredClass(suspendFullName))
-  private def makeMethodTypeWithContinuationType(t: Type, continuationType: Type)(
-      using ctx: Context): Type =
+  private def makeMethodTypeWithContinuationType(t: Type, continuationType: Type)(using ctx: Context): Type =
     val resultType = t.finalResultType
     val completionType = requiredClassRef(continuationFullName).appliedTo(continuationType)
     MethodType.apply(List(completionType), resultType)
@@ -18,20 +17,33 @@ final class ValDefAddCompletionType()(using Context) extends TypeMap {
     tp match {
       case t: TypeRef if hasSuspendClass(t.symbol) =>
         NoType
+      case t @ MethodType(termNames) =>
+        val resultType = t.finalResultType
+        val completionType = requiredClassRef(continuationFullName).appliedTo(resultType)
+        val completionParamIndex =
+          t.paramInfoss.indexWhere(_.exists(_ =:= requiredClassRef(suspendFullName)))
+        val withCompletion = t
+          .paramInfoss
+          .map(_.map(apply).filterNot(t => {
+            t == NoType
+          }))
+          .filterNot(_.isEmpty)
+          .insertAt(completionParamIndex, List(completionType))
+        withCompletion.foldRight(resultType)(MethodType.apply)
       case AppliedType(tycon, args) =>
         val argsWithoutSuspend = args.filterNot {
           case tr: TypeRef if hasSuspendClass(tr.symbol) =>
             true
           case _ => false
         }
-        val newType: Type =
+        val newType =
           if (argsWithoutSuspend.size == 1) argsWithoutSuspend.head
-          else if (tycon.isContextualMethod) {
-            val contextFunctionTypeName = s"scala.ContextFunction${argsWithoutSuspend.size}"
-            AppliedType(requiredClassRef(contextFunctionTypeName), argsWithoutSuspend)
-          } else AppliedType(tycon, argsWithoutSuspend)
-        val completionType = requiredClassRef(continuationFullName).appliedTo(newType)
-        MethodType.apply(List(completionType), newType)
+          else
+            if (tycon.isContextualMethod) {
+              val contextFunctionTypeName = s"scala.ContextFunction${argsWithoutSuspend.size}"
+              AppliedType(requiredClassRef(contextFunctionTypeName), argsWithoutSuspend)
+            } else AppliedType(tycon, argsWithoutSuspend)
+        newType
       case ExprType(resType) =>
         val withoutSuspend = apply(resType)
         val completionType = requiredClassRef(continuationFullName).appliedTo(withoutSuspend)
